Refactor of match.py to make it better readable (#76)
* refactored match.py

* now with new files
