+++ databallpy/__init__.py
@@ -1,12 +1,6 @@
 # read version from installed package
 from importlib.metadata import version
 
+from databallpy.get_match import get_match, get_open_match, get_saved_match
+
 __version__ = version("databallpy")
-
-
-class DataBallPyError(Exception):
-    "Error class specific for databallpy"
-
-    def __init__(self, message: str) -> None:
-        self.message = message
-        super().__init__(message)
