+++ setup.cfg
@@ -3,4 +3,5 @@ max-line-length = 88
 ignore = E203, W503
 per-file-ignores =
     tests/mocks.py: W191, E101
+    databallpy/__init__.py: F401
